Get things passing brad's validations
